[feat/#7] Fix number format exception

+++ composeApp/src/commonMain/kotlin/com/survivai/survivai/game/colosseum/ColosseumInfo.kt
@@ -168,7 +168,7 @@ object ColosseumInfo {
                     it.killPoint.toString(),
                     surviveTime.msToMMSS(),
                     it.maxComboPoint.toString(),
-                    score.toString(),
+                    score.toInt().toString(),
                 )
             }.sortedByDescending {
                 it.last().toInt()
